+++ lib/src/system/utils/matrix.dart
@@ -148,20 +148,7 @@ abstract class Matrix<T> {
   }
 
   @override
-  int get hashCode {
-    var result = 83;
-
-    // Like we did in operator== iterating over all elements ensures that the
-    // hashCode is properly calculated.
-    for (var i = 0; i < _data.length; ++i) {
-      result = result * 37 + _data[i].hashCode;
-    }
-
-    result = result * 37 + rowCount.hashCode;
-    result = result * 37 + columnCount.hashCode;
-
-    return result;
-  }
+  int get hashCode => Object.hashAll([rowCount, columnCount, ..._data]);
 
   @override
   String toString() {