+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - **BREAKING**: The `Complex.fromPolar` constructor now asks for required **named** parameters
  - **BREAKING**: The `Interpolation` constructor now asks for required **named** parameters
  - **BREAKING**: The `SylvesterMatrix` type now accepts an `Algebraic` type rather than a `List<Complex>`. As such, the `SylvesterMatrix.fromReal` constructor has been removed because a real polynomial can be built using `Algebraic.fromReal` instead
+ - **BREAKING**: The `SystemSolver` type now accepts a `RealMatrix` type rather than a `List<List<double>>`. As such, the `size` parameter has also been removed because the size can be retrieved from the `RealMatrix` object itself
  - Updated Dart SDK constraints to `">=2.17.0 <3.0.0"`
  - Added `csc` (cosecant) and `sec` (secant) trigonometric functions to the `ExpressionParser` type
  - Migrated Dart code to 2.17 with super parameters